Client: migrate styles to CSS modules
