+++ client/src/components/Field/Field.module.css
@@ -1,0 +1,59 @@
+.field {
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+}
+
+.label {
+  color: var(--color-slate);
+  font-size: 0.9rem;
+}
+
+.inputWrapper {
+  position: relative; 
+  width: 384px;
+}
+
+
+.input {
+  width: 100%;
+  height: 48px;
+  padding-left: 50px; 
+  border-radius: var(--border-radius);
+  border: var(--border);
+  color: var(--color-grey);
+  font-size: 1rem;
+}
+
+.input:focus {
+  outline: none;
+  border-color: var(--color-primary);
+}
+
+.icon {
+  position: absolute;
+  left: 15px;
+  top: 50%;
+  transform: translateY(-50%); 
+  color: var(--color-grey);
+  pointer-events: none;
+  font-size: 1.1rem;
+}
+
+.inputError {
+  background-color: var(--color-pink);
+  border-color: var(--color-red);
+  width: 100%;
+  height: 48px;
+  padding-left: 50px; 
+  border-radius: var(--border-radius);
+  border: var(--border);
+  color: var(--color-grey);
+  font-size: 1rem;
+}
+
+.error {
+  font-size: 0.8rem;
+  color: var(--color-red);
+  margin-left: 10px;
+}
