+++ client/src/components/Linking/Linking.module.css
@@ -1,0 +1,4 @@
+.Linking {
+	text-decoration: none;
+	color: var(--color-primary);
+}
