+++ client/src/components/Login/Login.module.css
@@ -1,0 +1,21 @@
+.login {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+	width: 100%;
+	max-width: 448px;
+	padding: 24px;
+	border: 1px solid black;
+	background-color: var(--color-white-solid);
+	border: var(--border);
+	border-radius: var(--border-radius);
+	gap: 25px;
+	box-shadow: 0 20px 40px rgba(0, 0, 0, 0.1), 0 8px 16px rgba(0, 0, 0, 0.05);
+}
+
+@media (max-width: 480px) {
+	.login {
+		padding: 20px;
+	}
+}
