+++ client/src/components/MainPage/ActiveState/ActiveState.jsx
@@ -1,0 +1,53 @@
+import DeleteAllButton from '../DeleteAllButton/DeleteAllButton'
+import TodoItem from '../ToDoItem/ToDoItem';
+import styles from './ActiveState.module.css'
+
+const ActiveState = ({
+	tasks,
+	setTasks,
+	onEdit,
+	activePomodoroId,
+	setActivePomodoroId,
+}) => {
+	const handleDeleteAll = () => {
+		setTasks((prev) => prev.filter((t) => t.done));
+	};
+
+	const toggleDone = (id) => {
+		setTasks((prev) =>
+			prev.map((t) => {
+				if (t.id === id && !t.done) {
+					setActivePomodoroId(null);
+				}
+				return t.id === id ? { ...t, done: !t.done } : t;
+			})
+		);
+	};
+
+	const deleteTask = (id) => {
+		setTasks((prev) => prev.filter((t) => t.id !== id));
+	};
+
+	return (
+		<div className={`${styles.activeState}`}>
+			<div className={`${styles.activeHeader}`}>
+				<h1 className={`${styles.activeStateCount}`}>Active tasks ({tasks.length})</h1>
+				<DeleteAllButton onClick={handleDeleteAll} />
+			</div>
+
+			{tasks.map((task) => (
+				<TodoItem
+					key={task.id}
+					task={task}
+					onToggle={toggleDone}
+					onDelete={deleteTask}
+					onEdit={onEdit}
+					activePomodoroId={activePomodoroId}
+					setActivePomodoroId={setActivePomodoroId}
+				/>
+			))}
+		</div>
+	);
+};
+
+export default ActiveState;
