+++ client/src/components/MainPage/ActiveState/ActiveState.module.css
@@ -1,0 +1,27 @@
+.activeState {
+  width: 100vw;
+  max-width: 864px;
+  min-height: 140px;
+  margin: 1rem 0;
+  padding: 1rem;
+  background-color: var(--color-white-solid);
+  border: var(--border);
+  border-radius: var(--border-radius);
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+}
+
+
+.activeHeader {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+}
+
+.activeStateCount {
+	margin: 0;
+	font-size: 18px;
+	font-weight: 400;
+	color: var(--color-dark-1);
+}
