+++ client/src/components/MainPage/AddTaskModal/AddTaskModal.module.css
@@ -1,0 +1,151 @@
+.modal {
+  position: fixed;
+  bottom: 24px;
+  right: 24px;
+  width: 380px;
+  background: var(--color-white-solid);
+  border: var(--border);
+  border-radius: var(--border-radius);
+  box-shadow: 0 10px 25px rgba(0,0,0,0.08);
+  animation: slideUp 0.25s ease forwards;
+  z-index: 50;
+}
+
+@keyframes slideUp {
+  from { transform: translateY(20px); opacity: 0; }
+  to   { transform: translateY(0); opacity: 1; }
+}
+
+.header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 14px 16px;
+  border-bottom: 1px solid #e5e7eb;
+  background: linear-gradient(to right, #eff6ff, #fff);
+}
+
+.title {
+  color: var(--color-dark-1);
+  font-weight: 400;
+  font-size: 18px;
+}
+
+.closeBtn {
+  padding: 4px;
+  border: none;
+  background: transparent;
+  color: #9ca3af;
+  cursor: pointer;
+  border-radius: 8px;
+  transition: 0.2s;
+}
+
+.closeBtn:hover {
+  background: #f3f4f6;
+  color: #4b5563;
+}
+
+.form {
+  padding: 16px;
+}
+
+.errorBox {
+  display: flex;
+  gap: 8px;
+  align-items: flex-start;
+  padding: 10px;
+  border-radius: 10px;
+  background: #fef2f2;
+  border: 1px solid #fecaca;
+  margin-bottom: 16px;
+}
+
+.errorIcon {
+  color: #dc2626;
+  margin-top: 2px;
+}
+
+.errorText {
+  color: #991b1b;
+  margin: 0;
+}
+
+.inputGroup label {
+  display: block;
+  color: #374151;
+  margin-bottom: 6px;
+}
+
+.input {
+  width: 100%;
+  padding: 12px;
+  border-radius: 12px;
+  border: 1px solid #e5e7eb;
+  outline: none;
+  transition: 0.2s;
+  font-size: 15px;
+}
+
+.input:focus {
+  border-color: #3b82f6;
+  box-shadow: 0 0 0 2px rgba(59,130,246,0.2);
+}
+
+.inputError {
+  border-color: #fca5a5 !important;
+  background: #fef2f2;
+}
+
+.inputErrorMsg {
+  margin-top: 6px;
+  color: #dc2626;
+}
+
+.footer {
+  display: flex;
+  justify-content: flex-end;
+  gap: 8px;
+  margin-top: 6px;
+}
+
+.btn {
+  border: none;
+  font-size: 14px;
+  padding: 10px 16px;
+  border-radius: 12px;
+  cursor: pointer;
+  transition: 0.2s;
+  display: flex;
+  align-items: center;
+  gap: 6px;
+}
+
+.cancel {
+  background: transparent;
+  color: #374151;
+}
+
+.cancel:hover {
+  background: #f3f4f6;
+}
+
+.add {
+  background: #2563eb;
+  color: white;
+}
+
+.add:hover {
+  background: #1d4ed8;
+}
+
+@media (max-width: 480px) {
+  .modal {
+    left: 0;
+    right: 0;
+    bottom: 24px;
+    margin: 0 auto;
+    width: min(380px, calc(100% - 32px));
+    transform: none;
+  }
+}
