--- client/src/components/MainPage/CreateTaskButton.jsx
@@ -1,15 +1,0 @@
-import '../../styles/elements/button.css';
-
-const CreateTaskButton = ({ inner, onClick }) => {
-	const handleClick = (e) => {
-		if (onClick) onClick(e);
-	};
-
-	return (
-		<button className='button' onClick={handleClick}>
-			{inner}
-		</button>
-	);
-};
-
-export default CreateTaskButton;
