+++ client/src/components/MainPage/DeleteAllButton/DeleteAllButton.module.css
@@ -1,0 +1,14 @@
+.deleteAllButton {
+	display: flex;
+	align-items: center;
+	gap: 0.5rem;
+	font-weight: 400;
+	background-color: var(--color-red);
+	border: none;
+	border-radius: var(--border-radius);
+	color: var(--color-white-solid);
+	cursor: pointer;
+	font-weight: 500;
+	font-size: 14px;
+	padding: 0.6rem 1rem;
+}
