+++ client/src/components/MainPage/EmptyState/EmptyState.module.css
@@ -1,0 +1,29 @@
+.emptyState {
+  width: 100vw;
+  max-width: 864px;
+  min-height: 140px;
+  margin: 1rem 0;
+  padding: 1rem;
+  background-color: var(--color-white-solid);
+  border: var(--border);
+  border-radius: var(--border-radius);
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+}
+
+.emptyStateCount {
+  margin: 0;
+  text-align: left;
+  font-size: 18px;
+  font-weight: 400;
+  color: var(--color-dark-1);
+}
+
+.emptyStateText {
+  margin: 0;
+  text-align: center;
+  color: var(--color-grey);
+  font-size: 1rem;
+  line-height: 1.4;
+}
