--- client/src/components/MainPage/MainPage.css
@@ -1,52 +1,0 @@
-.main-content {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	width: 100%;
-	padding: 1rem 0rem;
-}
-
-.todo-container {
-	width: 100%;
-	max-width: 864px;
-	display: flex;
-	flex-direction: column;
-}
-
-.delete-all-button {
-	display: flex;
-	align-items: center;
-	gap: 0.5rem;
-	font-weight: 400;
-	background-color: var(--color-red);
-	border: none;
-	border-radius: var(--border-radius);
-	color: var(--color-white-solid);
-	cursor: pointer;
-	font-weight: 500;
-	font-size: 14px;
-	padding: 0.6rem 1rem;
-}
-
-.delete-all-button:hover {
-	background-color: #ce0000;
-}
-
-.active-header {
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-}
-
-.active-header h1 {
-	margin: 0;
-	font-size: 5px;
-	font-weight: 400;
-	color: var(--color-dark-1);
-}
-
-.completed-list {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
