+++ client/src/components/Navbar/Navbar.module.css
@@ -1,0 +1,103 @@
+.nav {
+	position: fixed;
+	top: 0;
+	left: 0;
+	width: 100%;
+	height: 93px;
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	background-color: #fff;
+	border-bottom: 1px solid #eaeaea;
+	z-index: 1000;
+	padding: 0 16px;
+}
+
+.inner {
+	width: 100%;
+	max-width: 874px;
+	margin: 0 auto;
+	padding: 0 16px;
+
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
+}
+
+.left {
+	display: flex;
+	align-items: center;
+	gap: 12px;
+}
+
+.backButton {
+    background: none;
+    border: none;
+    width: 38px;
+    height: 38px;
+    border-radius: 50%;
+    font-size: 20px;
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.backButton:hover {
+	background-color: var(--color-grey-light);
+}
+
+.titleEmail {
+	display: flex;
+	flex-direction: column;
+}
+
+.title {
+	font-size: 18px;
+	font-weight: 600;
+	line-height: 1.2;
+}
+
+.userEmail {
+	font-size: 13px;
+	color: #666;
+	line-height: 1.2;
+}
+
+.right {
+	display: flex;
+	align-items: center;
+	gap: 12px;
+	
+}
+
+.icon {
+	background: none;
+	border: none;
+	border-radius: 50%;
+	cursor: pointer;
+	font-size: 20px;
+	padding: 8px;
+	line-height: 1;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+}
+
+.icon:hover {
+	background-color: var(--color-grey-light);
+}
+
+@media (min-width: 600px) {
+  .left {
+    gap: 6px;
+  }
+
+  .appName {
+    font-size: 20px;
+  }
+
+  .userEmail {
+    font-size: 14px;
+  }
+}
