--- client/src/styles/elements/active-tasks.css
@@ -1,63 +1,0 @@
-.content {
-	display: flex;
-	justify-content: center;
-	width: 100%;
-	margin-top: 20px;
-}
-
-.container {
-	width: 100%;
-	max-width: 864px;
-	margin: 0 auto;
-}
-
-.box {
-  width: 100%;
-  padding: 24px;
-  box-sizing: border-box; 
-}
-
-.box-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.box-title {
-  font-size: 18px;
-}
-
-.delete-all {
-  background-color: var(--color-red);
-  color: var(--color-white-solid);
-  border: none;
-  padding: 8px 16px;
-  border-radius: var(--border-radius);
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  cursor: pointer;
-  transition: all .4s;
-}
-
-.delete-all:hover {
-  background-color: var(--color-red-hover);
-}
-
-
-@media (max-width: 768px) {
-  .box {
-    padding: 16px;
-    border-radius: 8px;
-  }
-
-  .box-title {
-    font-size: 16px;
-
-  }
-
-  .delete-all {
-    padding: 6px 12px;
-    font-size: 14px;
-  }
-}
