--- client/src/styles/elements/completed-tasks.css
@@ -1,75 +1,0 @@
-.main-page-content {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  width: 100%;
-  gap: 24px; 
-  margin-top: 20px;
-  
-}
-
-
-.main-page-container {
-  width: 100%;
-  max-width: 864px;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-}
-
-.main-page-box {
-    width: 100%;
-    background-color: var(--color-white-solid);
-    padding: 24px;
-    border-radius: 12px;
-    box-shadow: 0 2px 8px rgba(0,0,0,0.05);
-    display: flex;
-    flex-direction: column;
-    gap: 16px;
-}
-
-.box-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.box-title {
-  font-size: 18px;
-}
-
-.delete-all {
-  background-color: var(--color-red);
-  color: var(--color-white-solid);
-  border: none;
-  padding: 8px 16px;
-  border-radius: var(--border-radius);
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  cursor: pointer;
-  transition: all .4s;
-}
-
-.delete-all:hover {
-  background-color: #c9000a;
-}
-
-
-@media (max-width: 768px) {
-  .main-page-box {
-    padding: 16px;
-    border-radius: 8px;
-  }
-
-  .box-title {
-    font-size: 16px;
-
-  }
-
-  .delete-all {
-    padding: 6px 12px;
-    font-size: 14px;
-  }
-}
