--- client/src/styles/elements/field.css
@@ -1,44 +1,0 @@
-.field {
-	position: relative;
-}
-
-.field input {
-	display: flex;
-	flex-direction: row;
-	align-items: flex-start;
-	width: 384px;
-	height: 48px;
-	color: var(--color-grey);
-	padding-left: 50px;
-	border-radius: var(--border-radius);
-	border: var(--border);
-}
-
-.field input:focus {
-	outline: none;
-}
-
-.field label {
-	color: var(--color-slate);
-	pointer-events: none;
-}
-
-.icon {
-	position: absolute;
-	left: 15px;
-	top: 50%;
-	color: var(--color-grey);
-	pointer-events: none;
-}
-
-.input--error {
-	background: var(--color-pink);
-}
-
-.field-error {
-	position: absolute;
-	left: 10px;
-	top: 73px;
-	color: var(--color-red);
-	font-size: 0.8rem;
-}
