--- client/src/styles/elements/heading.css
@@ -1,16 +1,0 @@
-.heading__title {
-	font-weight: 500;
-	font-size: 22px;
-	line-height: 160%;
-	text-align: center;
-	color: var(--color-dark-1);
-	padding-bottom: 10px;
-}
-
-.heading__text {
-	font-weight: 400;
-	font-size: 15px;
-	line-height: 160%;
-	text-align: center;
-	color: var(--color-dark-2);
-}
