--- client/src/styles/elements/linking.css
@@ -1,4 +1,0 @@
-.link {
-	text-decoration: none;
-	color: var(--color-primary);
-}
