--- client/src/styles/elements/navbar.css
@@ -1,104 +1,0 @@
-.nav {
-	position: fixed; /* przykleja do góry */
-	top: 0;
-	left: 0;
-	width: 100%;
-	height: 93px;
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	background-color: #fff;
-	border-bottom: 1px solid #eaeaea;
-	z-index: 1000; /* żeby była nad contentem */
-	padding: 0 16px;
-}
-
-.nav__inner {
-	width: 100%;
-	max-width: 874px;
-	margin: 0 auto;
-	padding: 0 16px;
-
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-}
-
-.nav__left {
-	display: flex;
-	align-items: center;
-	gap: 12px;
-}
-
-.back-button {
-    background: none;
-    border: none;
-    width: 38px;
-    height: 38px;
-    border-radius: 50%;
-    font-size: 20px;
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.back-button:hover {
-	background-color: var(--color-grey-light);
-}
-
-.title-email {
-	display: flex;
-	flex-direction: column;
-}
-
-.nav__title {
-	font-size: 18px;
-	font-weight: 600;
-	line-height: 1.2;
-}
-
-.nav__user-email {
-	font-size: 13px;
-	color: #666;
-	line-height: 1.2;
-}
-
-.nav__right {
-	display: flex;
-	align-items: center;
-	gap: 12px;
-	
-}
-
-.nav__icon {
-	background: none;
-	border: none;
-	border-radius: 50%;
-	cursor: pointer;
-	font-size: 20px;
-	padding: 8px;
-	line-height: 1;
-
-	display: flex;
-	align-items: center;
-	justify-content: center;
-}
-
-.nav__icon:hover {
-	background-color: var(--color-grey-light);
-}
-
-@media (min-width: 600px) {
-	.nav__left {
-		gap: 6px;
-	}
-
-	.nav__app-name {
-		font-size: 20px;
-	}
-
-	.nav__user-email {
-		font-size: 14px;
-	}
-}
